+++ lib/augmented/exceptions.rb
@@ -1,7 +1,9 @@
+require 'augmented/exceptions/chain'
 require 'augmented/exceptions/detailed'
 
 module Augmented
   module Exceptions
+    include Chain
     include Detailed
   end
 end