+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Added `Exception#chain`
 - Added `Exception#details`, `Exception#details=`, `Exception#detailed`
 
 ## [0.2.3] - 2021-05-29